backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
pydantic[email]
python-multipart
